Adjazenz Array benutbar in Graphen

Closes #40

See merge request spuehlem/semesterprojekt_ws2425!49

Former-commit-id: ae5f38140f0bfc4b3be0740bddc18847bacda5da

+++ src/AttributedDeepwalk.cpp
@@ -245,7 +245,8 @@ double AttributedDeepwalk::measuring_structural_similarity(int node1, int node2)
         return 0.0;
     }
 
-    return intersectionOfCovers.size() / unionOfCovers.size();
+    return static_cast<double>(intersectionOfCovers.size()) / unionOfCovers.size();
+
 }
 
 /*
